test(Form.SubmitConfirmation): colocate listener timing coverage
